# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by specific date (YYYY-MM-DD)
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, completed, cancelled, no_show]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Billing ───────────────────────────────────────────────────────────────
  /bills:
    get:
      operationId: listBills
      tags: [billing]
      summary: List bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partial, paid, overdue]
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [billing]
      summary: Create a bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
  /bills/{id}:
    get:
      operationId: getBill
      tags: [billing]
      summary: Get a bill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
    patch:
      operationId: updateBill
      tags: [billing]
      summary: Update a bill (e.g. record payment, change status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
    delete:
      operationId: deleteBill
      tags: [billing]
      summary: Delete a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Inventory ─────────────────────────────────────────────
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/Employee"

    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found

    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
      responses:
        "204":
          description: Deleted

  
  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, patientId, patientName, dentistName, appointmentDate, appointmentTime, durationMinutes, type, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        dentistName:
          type: string
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          description: "HH:MM format"
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, root_canal, crown, whitening, orthodontics, xray, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [patientId, dentistName, appointmentDate, appointmentTime, durationMinutes, type]
      properties:
        patientId:
          type: integer
        dentistName:
          type: string
          minLength: 1
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, root_canal, crown, whitening, orthodontics, xray, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        dentistName:
          type: string
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, root_canal, crown, whitening, orthodontics, xray, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]

    # ── Bill ─────────────────────────────────────────────────────────────────
    Bill:
      type: object
      required: [id, patientId, patientName, totalAmount, amountPaid, status, dueDate, items, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        appointmentId:
          type: ["integer", "null"]
        totalAmount:
          type: number
        amountPaid:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue]
        dueDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillLineItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BillLineItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
    BillInput:
      type: object
      required: [patientId, totalAmount, dueDate, items]
      properties:
        patientId:
          type: integer
        appointmentId:
          type: integer
        totalAmount:
          type: number
        amountPaid:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue]
        dueDate:
          type: string
          format: date
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillLineItem"
    BillUpdate:
      type: object
      properties:
        totalAmount:
          type: number
        amountPaid:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue]
        dueDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillLineItem"

    # ── Inventory ─────────────────────────────────────────────────────────────
...
    Employee:
      type: object
      required: [id, fullName, role, email, phone, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [fullName, role, email, phone]
      properties:
        fullName:
          type: string
          minLength: 1
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
              

         

    # ── Dashboard ─────────────────────────────────────────────────────────────
...