# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';
import type { AppointmentUpdateType } from './appointmentUpdateType';

export interface AppointmentUpdate {
  dentistName?: string;
  appointmentDate?: Date;
  appointmentTime?: string;
  durationMinutes?: number;
  type?: AppointmentUpdateType;
  status?: AppointmentUpdateStatus;
  /** @nullable */
  notes?: string | null;
}