# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BillLineItem } from './billLineItem';
import type { BillStatus } from './billStatus';

export interface Bill {
  id: number;
  patientId: number;
  patientName: string;
  /** @nullable */
  appointmentId?: number | null;
  totalAmount: number;
  amountPaid: number;
  status: BillStatus;
  dueDate: Date;
  /** @nullable */
  notes?: string | null;
  items: BillLineItem[];
  createdAt: Date;
  updatedAt: Date;
}