# lib/api-zod/src/generated/types/billUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BillLineItem } from './billLineItem';
import type { BillUpdateStatus } from './billUpdateStatus';

export interface BillUpdate {
  totalAmount?: number;
  amountPaid?: number;
  status?: BillUpdateStatus;
  dueDate?: Date;
  /** @nullable */
  notes?: string | null;
  items?: BillLineItem[];
}